Extract modal.styles.js
