+++ src/components/wallet/invalid.rskaddress.confirmation.js
@@ -11,89 +11,9 @@ import config from '../../../config';
 import common from '../../common/common';
 import Checkbox from '../common/misc/checkbox';
 import { strings } from '../../common/i18n';
+import modalStyles from '../../assets/styles/modal.styles';
 
 const styles = StyleSheet.create({
-  container: {
-    backgroundColor: color.blackA50,
-    justifyContent: 'center',
-    flex: 1,
-  },
-  panel: {
-    backgroundColor: color.white,
-    borderRadius: 5,
-    marginHorizontal: 25,
-    maxHeight: '82%',
-    paddingHorizontal: 25,
-    paddingBottom: 20,
-  },
-  title: {
-    fontSize: 17,
-    color: color.black,
-    fontFamily: fontFamily.AvenirHeavy,
-    textAlign: 'center',
-  },
-  message: {
-    fontSize: 14,
-    color: color.black,
-    fontFamily: fontFamily.AvenirBook,
-  },
-  link: {
-    fontSize: 14,
-    color: color.app.theme,
-    fontFamily: fontFamily.AvenirBook,
-  },
-  rowContainer: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-  },
-  notice: {
-    fontSize: 16,
-    color: color.black,
-    marginTop: 10,
-    marginBottom: 20,
-  },
-  subdomain: {
-    fontSize: 16,
-    color: color.black,
-    fontFamily: fontFamily.AvenirBlack,
-  },
-  subdomainRow: {
-    alignItems: 'center',
-  },
-  recommendButton: {
-    justifyContent: 'center',
-    marginTop: 25,
-    height: 45,
-    backgroundColor: color.app.theme,
-    borderRadius: 20,
-    width: '90%',
-    alignItems: 'center',
-  },
-  normalButton: {
-    justifyContent: 'center',
-    height: 45,
-    borderRadius: 20,
-    width: '90%',
-    alignItems: 'center',
-    borderColor: color.vividBlue,
-    borderWidth: 1,
-  },
-  disabledButton: {
-    borderColor: color.gray91,
-  },
-  recommendButtonText: {
-    color: color.concrete,
-  },
-  normalButtonText: {
-    fontFamily: fontFamily.AvenirHeavy,
-    color: color.vividBlue,
-    fontSize: 16,
-  },
-  disabledButtonText: {
-    fontFamily: fontFamily.AvenirHeavy,
-    color: color.gray91,
-    fontSize: 16,
-  },
   row: {
     flexDirection: 'row',
     marginTop: 10,
@@ -111,10 +31,6 @@ const styles = StyleSheet.create({
   },
   address: {
     fontSize: 12,
-  },
-  buttonsView: {
-    marginTop: 23,
-    alignItems: 'center',
   },
 });
 
@@ -189,12 +105,12 @@ class InvalidRskAddressConfirmation extends Component {
                 <Checkbox text={strings('modal.invalidRskAddress.risk')} isChecked={isRiskConfirmed} onValueChanged={this.onRiskValueChanged} />
               </View>
             </View>
-            <View style={[styles.buttonsView]}>
-              <TouchableOpacity style={[styles.normalButton, isRiskConfirmed ? null : styles.disabledButton]} disabled={!isRiskConfirmed} onPress={this.onConfirmPressed}>
-                <Loc style={[styles.normalButtonText, isRiskConfirmed ? null : styles.disabledButtonText]} text="modal.invalidRskAddress.confirm" />
+            <View style={[modalStyles.buttonsView]}>
+              <TouchableOpacity style={[modalStyles.normalButton, isRiskConfirmed ? null : modalStyles.disabledButton]} disabled={!isRiskConfirmed} onPress={this.onConfirmPressed}>
+                <Loc style={[modalStyles.normalButtonText, isRiskConfirmed ? null : modalStyles.disabledButtonText]} text="modal.invalidRskAddress.confirm" />
               </TouchableOpacity>
-              <TouchableOpacity style={[styles.recommendButton, space.marginTop_10]} onPress={this.onCancelPressed}>
-                <Loc style={[styles.normalButtonText, styles.recommendButtonText]} text="modal.invalidRskAddress.cancel" />
+              <TouchableOpacity style={[modalStyles.recommendButton, space.marginTop_10]} onPress={this.onCancelPressed}>
+                <Loc style={[modalStyles.normalButtonText, styles.recommendButtonText]} text="modal.invalidRskAddress.cancel" />
               </TouchableOpacity>
             </View>
           </View>
